fix: reusable timeSince util
